+++ src/composables/common/useResizablePanel.ts
